convert executables from esm to cjs
