+++ .github/actions/compare-base/action.yml
@@ -48,7 +48,7 @@ runs:
         npm ci --ignore-scripts
         rm -f "$JHI_BIN_FOLDER/jhipster"
         npm uninstall -g $JHI_CLI_PACKAGE
-        ln -s "${{ github.workspace }}/base/$JHI_CLI_PACKAGE/bin/jhipster.mjs" "$JHI_BIN_FOLDER/jhipster"
+        ln -s "${{ github.workspace }}/base/$JHI_CLI_PACKAGE/bin/jhipster.cjs" "$JHI_BIN_FOLDER/jhipster"
         jhipster --install-path
       shell: bash
     - name: 'MERGE: merge base config'
