+++ generators/generate-blueprint/generator.mjs
@@ -323,13 +323,15 @@ export default class extends BaseGenerator {
         } = this.options;
         if (!generateSnapshots) return;
 
-        // Generate snapshots to add to git.
-        this.logger.info(`
-This is a new blueprint, executing '${chalk.yellow('npm run update-snapshot')}' to generate snapshots and commit to git.`);
         try {
           if (this.options[LINK_JHIPSTER_DEPENDENCY]) {
-            await this.spawnCommand('npm', ['link', 'generator-jhipster']);
+            this.logger.info('Linking generator-jhipster');
+            await this.spawnCommand('npm', ['link', 'generator-jhipster'], { stdio: 'inherit' });
           }
+
+          // Generate snapshots to add to git.
+          this.logger.info(`
+This is a new blueprint, executing '${chalk.yellow('npm run update-snapshot')}' to generate snapshots and commit to git.`);
           await this.spawnCommand('npm', ['run', 'update-snapshot']);
         } catch (error) {
           if (generateSnapshots !== undefined) {
